requirements.txt
streamlit==1.52.1
langchain==1.1.2
langchain-openai==1.1.0
langchain-core==1.1.1
langgraph==1.0.4
langchain-openai==1.1.0
yt-dlp==2025.11.12
openai-whisper==20250625
librosa==0.11.0
numpy==2.2.6
soundfile==0.13.1
pillow==12.0.0
ffmpeg-python==0.2.0
resampy==0.4.3
opencv-python==4.12.0.88

requests
python-dotenv